bugfix: make pylint love me again
