+++ scripts/cat/factories/create_cat.py
@@ -2,9 +2,11 @@ from __future__ import annotations
 
 from random import randint, sample, choice
 
-from scripts.cat.enums import CatRank
+from scripts.cat.cats import Cat
+from scripts.cat.enums import CatRank, CatAge
 from scripts.cat.factories.cat_factory import CatFactory
 from scripts.cat.factories.enums import CatType
+from scripts.cat.pelts import Pelt
 from scripts.game_structure import game
 
 
@@ -62,3 +64,35 @@ def create_example_cats():
             game.choose_cats[cat_index] = create_cat(
                 cat_type=CatType.NEW, rank=random_rank
             )
+
+
+def create_option_preview_cat(scar: str = None, acc: str = None):
+    """
+    Creates a cat with the specified scar
+    """
+    new_cat = CatFactory.create_cat(
+        CatType.TEST,
+        loading_cat=True,
+        pelt=Pelt(
+            name="SingleColour",
+            colour="WHITE",
+            length="medium",
+            eye_color="SAGE",
+            reverse=False,
+            white_patches=None,
+            vitiligo=None,
+            points=None,
+            tortie_marking=None,
+            tortie_base=None,
+            tortie_pattern=None,
+            tortie_colour=None,
+            tint="gray",
+            skin="BLUE",
+            scars=[scar] if scar else [],
+            adult_sprite="8",
+            accessory=[acc] if acc else [],
+        ),
+    )
+    new_cat.age = CatAge.ADULT
+
+    return new_cat
